# docs/agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Studio",
    "version": "1.0.3",
    "description": "Agent-facing API. Branch on the `code` field of an error, never on its prose.\nIn server mode every call carries `Authorization: Bearer avs_…`; scopes are read, produce, publish, admin.\nName the channel explicitly (body.channelId, ?channel= or X-AVS-Channel) — the active channel is only a default.\nSend an Idempotency-Key (or clientRef on a create) so a retry after a timeout cannot spend twice.\nPoll GET /events?after=<id> for progress; a WebSocket at /ws carries the same feed live."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "error"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine word. Branch on this, never on `error`."
          },
          "error": {
            "type": "string",
            "description": "Short reason, in the user's language unless ?lang= says otherwise."
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        }
      },
...
      "ProjectSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "topic": {
            "type": "string",
            "description": "First 200 characters only in list replies."
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "running",
              "paused",
              "scenes",
              "review",
              "done",
              "error"
            ]
          },
          "current_step": {
            "type": "string",
            "nullable": true,
            "description": "b2|b34|b5|b6|b7"
          },
          "aspect_ratio": {
            "type": "string"
          },
          "channel_id": {
            "type": "string"
          },
          "video_path": {
            "type": "string",
            "nullable": true
          },
          "thumb_path": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "scenes_approved_at": {
            "type": "integer",
            "nullable": true
          },
          "created_at": {
            "type": "integer"
          },
          "updated_at": {
            "type": "integer"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Mint with `npm run token -- create` on the machine running the server."
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/projects": {
      "get": {
        "summary": "Project summaries for a channel, newest first.",
        "tags": [
          "projects"
        ],
        "x-avs-scope": "read",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Channel id. Also accepted as X-AVS-Channel or body.channelId. Defaults to the token’s channel, then the active one."
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "short",
                "landscape"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a project. Send clientRef (or an Idempotency-Key header) so a retry cannot create a second one.",
        "tags": [
          "projects"
        ],
        "x-avs-scope": "produce",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "topic"
                ],
                "properties": {
                  "topic": {
                    "type": "string"
                  },
                  "channelId": {
                    "type": "string"
                  },
                  "clientRef": {
                    "type": "string",
                    "description": "Your own reference. Unique per channel; a repeat returns the first project with reused:true."
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Overrides layered over channel defaults and the default preset."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/ProjectSummary"
                    },
                    "reused": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Known codes: channel_not_found, channel_denied, idempotency_key_reused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "summary": "One project. ?scenes=lite drops the generated page from every scene; ?scenes=0 returns the project alone.",
        "tags": [
          "projects"
        ],
        "x-avs-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scenes",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "full",
                "lite",
                "0"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Known codes: not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Edit title, topic, aspect ratio, config or metadata.",
        "tags": [
          "projects"
        ],
        "x-avs-scope": "produce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Known codes: not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a project and its files.",
        "tags": [
          "projects"
        ],
        "x-avs-scope": "produce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Known codes: not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{id}/restart": {
      "post": {
        "summary": "Start again as a NEW project; the previous attempt survives for comparison.",
        "tags": [
          "pipeline"
        ],
        "x-avs-scope": "produce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Known codes: not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{id}/qc-scan": {
      "get": {
        "summary": "Read the finished project back and report what a person would not catch.",
        "tags": [
          "quality"
        ],
        "x-avs-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/typeset-scan": {
      "get": {
        "summary": "Scenes whose text is at risk of overflowing its frame.",
        "tags": [
          "quality"
        ],
        "x-avs-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}